\name{HCD-package}
\alias{HCD-package}
\docType{package}
\title{
  Hierarchical community detection by recursive partitioning
}
\description{
The package provides the implementation of the recursive partitioning
strategy to clustering network nodes in a hierarchical way. It also
includes the mechanism of generating networks from a binary tree
stochastic block model.
}
\details{
\tabular{ll}{
Package: \tab HCD\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2024-01-28\cr
License: \tab GPL (>= 2)\cr
}

}
\author{
  Tianxi Li, Lihua Lei, Sharmodeep Bhattacharyya, Purnamrita Sarkar, Peter Bickel, and Elizaveta Levina.\cr

Maintainer: Tianxi Li <tianxili@umn.edu>
}
\references{
Li, T., Lei, L., Bhattacharyya, S., Van den Berge, K., Sarkar, P., Bickel, P.J. and Levina, E., 2022. Hierarchical community detection by recursive partitioning. Journal of the American Statistical Association, 117(538), pp.951-968.
}
\keyword{ package }
