\name{HCD-package}
\alias{HCD-package}
\docType{package}
\title{
  Hierarchical community detection by recursive partitioning
}
\description{
The package provides the implementation of the recursive partitioning
strategy to clustering network nodes in a hierarchical way. It also
includes the mechanism of generating networks from a binary tree
stochastic block model.
}
\details{
\tabular{ll}{
Package: \tab HCD\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2019-08-02\cr
License: \tab GPL (>= 2)\cr
}

}
\author{
  Tianxi Li, Lihua Lei, Sharmodeep Bhattacharyya, Purnamrita Sarkar, Peter Bickel, and Elizaveta Levina.\cr

Maintainer: Tianxi Li <tianxili@virginia.edu>
}
\references{
Tianxi Li, Lihua Lei, Sharmodeep Bhattacharyya, Purnamrita Sarkar, Peter
Bickel, and Elizaveta Levina.  Hierarchical community detection by
recursive partitioning. arXiv:1810.01509
}
\keyword{ package }
